Separate EF Context scope from WriteLock scope
Since the EF DbContext isn't thread safe, we can't support having
multiple simultaneous operations on the context. However, we need
a single instance of the write lock (otherwise the lock is pretty
useless).

This makes it possible to register the write lock as singleton,
and the EF context as per request (or even transient).