Add a rudimentary event EventRegistry
This paves the way for adding transforms of events, e.g. after a refactoring
that renamed one of the event types.

It will also make it possible to build registries that look up types and names
without having to use the assembly qualified name (e.g. by specifying the
assembly to look in when configuring the registry) or even with custom naming
schemes.