Make the event registry mechanism composable
The event registry is now built manually from a few simple pieces,
instead of incrementally by a builder. This makes the implementation
a lot easier to work with, but with a little extra overhead in
configuration.

Perhaps, if there's demand for it, the builder pattern can be re-
introduced later to simplify config.