Separate the concepts of an event, and the storage type for it
Before, it was pretty awkward to get a hold of both the metadata of an
event (timestamp, version etc) and the actual payload, because the payload
was stored in serialized form on the metadata wrapper.

With this change, the storage type of the event is separated from the
metadata container, making it easier to create state projections that
use the metadata as well as the payload.