Update xUnit and Moq
Due to improved support for parallel test runs in xUnit 2.3, we need to
explicitly disable parallel test runs of the EF tests, because they're
all re-using the same db instance.

A better solution would be to refactor the test fixture and base class
to create a new db-connection instance for each test, and use that.