Don't expose the version number (#47)
Exposing the version number is problematic, since if we want to change
history by deleting or adding events in the past, we have to re-version
every single event that comes later in the stream.

Instead, we now use timestamps for versioning.

However, neither EF nor EF Core guarantees the ordering of entities that
are inserted in the same batch, which means we need a way to version
events that are inserted togehter to ensure that they are retreived in
the same order they were created. Therefore, we keep the Version property
on the persisted event types, but use it only internally inside the EF
providers.

The version number is also now grouped per timestamp, so that for each new
timestamp the version number restarts.